Add DACVAEEncoder and migrate MLSDataset to HuggingFace
- encoders/dac_vae.py: DAC encoder + trainable VAE bottleneck (새 인코더)
- encoders/__init__.py: DACVAEEncoder 등록 (key: "dac_vae")
- dataset.py: MLSDataset을 parler-tts/mls_eng_10k (HuggingFace)으로 교체;
  io.BytesIO + torchaudio.load로 직접 디코딩
- model.py: fp16 → bfloat16 (Qwen3.5-4B fp16 NaN 문제 대응)
- train.py, run.sh: dac_vae 옵션 추가
- docs/dac_vs_dac_vae.md: DAC vs DAC-VAE 비교 문서

+++ model.py
@@ -31,8 +31,8 @@ class AudioQwen(nn.Module):
         self.encoder = encoder
         self._cfg    = cfg  # apply_lora에서 LoRA 설정 참조
 
-        # V100: bf16은 소프트웨어 에뮬레이션만 지원, cuDNN LSTM은 거부 → fp16 고정
-        torch_dtype = torch.float16
+        # Qwen3.5-4B는 SSM 아키텍처로 fp16에서 NaN 발생 → bf16 사용
+        torch_dtype = torch.bfloat16
         cache_dir   = cfg["model_cache_dir"]
         llm_name    = cfg["llm_model"]
